speed sampling added

+++ src/Event.cpp
@@ -8,13 +8,13 @@
 
 Event Event::NO_EVENT(NoEvent, 0);
 
-Event::Event(EventType eventType, int eventData): type(eventType), data(eventData){};
+Event::Event(EventType eventType, uint8_t eventData): type(eventType), data(eventData){};
 
 EventType Event::getType() {
 	return type;
 }
 
-int Event::getData() {
+uint8_t Event::getData() {
 	return data;
 }
